+++ src/apprunner/wizards/imageRepositoryWizard.ts
@@ -22,7 +22,7 @@ import { WizardPrompter } from '../../shared/ui/wizardPrompter'
 import * as nls from 'vscode-nls'
 import { WizardForm } from '../../shared/wizards/wizardForm'
 import { createVariablesPrompter } from '../../shared/ui/common/variablesPrompter'
-import { makeDeploymentButton } from './apprunnerCreateServiceWizard'
+import { makeDeploymentButton } from './deploymentButton'
 import { IamClient } from '../../shared/clients/iamClient'
 import { RolePrompter } from '../../shared/ui/common/rolePrompter'
 import { getLogger } from '../../shared/logger/logger'
